Remove the remote authenticator
